# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "StickerHub API",
    "version": "0.1.0",
    "description": "API para el album de stickers virtual del Mundial 2026. Colecciona stickers, abre sobres, e intercambia con otros usuarios."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Desarrollo local"
    }
  ],
  "paths": {
...
    "/api/trades/{id}/accept": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": { "type": "integer" },
          "description": "ID de la oferta de intercambio"
        }
      ],
      "put": {
        "summary": "Aceptar una oferta de intercambio",
        "tags": ["Trades"],
        "security": [{ "cookieAuth": [] }],
        "responses": {
          "200": { "description": "Intercambio completado exitosamente" },
          "401": { "description": "No autorizado" },
          "404": { "description": "Oferta no encontrada o ya no esta pendiente" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sb-wsqbrbucfrmwyzajeyoi-auth-token",
        "description": "Session cookie de Supabase"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Token JWT de Supabase (usado en POST /api/listings)"
      }
    }
  },
...
}